fix(#1/#2): triage test_misc_op — arctan2 float16 promotion + CTCLoss export + test API
Triaged the 7 pre-existing test_misc_op failures (7 -> 5):
- test_arctan2 (REAL): arctan2 did `bool_mask * np.pi`, which under the torch dtype lattice
  promotes bool x python-float to float16 and rounds pi to 3.140625, breaking atol=1e-6.
  Fixed misc.py: cast the mask to the angle's float dtype before scaling by pi. (The broader
  bool x python-float -> float16 vs torch's float32 is a separate promotion fix, next commit.)
- test_multinorm (test API drift): jt.float32([...]) is now a dtype object, not a constructor;
  switched the test to jt.array(..., dtype="float32"). multinomial itself was correct.
- test_ctc_loss (partial): nn had no CTCLoss; re-export CTCLoss/ctc_loss from jittor.misc so
  nn.CTCLoss (and the shim's torch.nn.CTCLoss) resolve.

Remaining 5 (report-only, not jittor compute bugs): test_code_softmax + test_ctc_loss both
hit a torch_compat _grad_compat bug (jt.grad(loss, single_Var) returns a list not a Var —
fixed in a follow-up); test_make_grid x3 need real torchvision (jt311 ships the jittor stub).

Verified cscg104: test_misc_op 26 tests, 5 failures (was 7), no regressions.

+++ python/jittor/misc.py
@@ -809,9 +809,14 @@ def arctan2(y,x):
     x = (x!=0.0).ternary(x, 1e-30)
     angle = (y/x).arctan()
     mask = (x<0)&(y<0)
-    angle = angle - mask*np.pi
+    # mask is bool; `bool * python-float` promotes to float16 under the torch
+    # dtype lattice, rounding pi to 3.14 (3.140625 after upcast) and breaking
+    # the atol=1e-6 contract. Cast the mask to the angle's float dtype first so
+    # pi keeps full precision (float32 -> 3.1415927) while still honoring a
+    # genuinely float16/float64 input.
+    angle = angle - mask.cast(angle.dtype)*np.pi
     mask = (x<0)&(y>=0)
-    angle = angle + mask*np.pi
+    angle = angle + mask.cast(angle.dtype)*np.pi
     return angle
 atan2 = arctan2
